[MRG] Add support for random "base estimator" in Optimizer (#418)
* Add support for random "base estimator" in Optimizer

This removes the need to special case dummy_minimize by adding
the ability to switch to supplying only random suggestions in
the Optimizer class.

* Prevent gradients from going to zero

* Switch parameter name to "dummy"

* Fix GBRT quantile regressor predict() output

The GBRT regressor returns an array of predictions for each sample
containing the value of each of the quantiles. All other regressors
return just the mean. Fixed the GBRT regressor to behave like the
others.

* Make if statement structure clearer

* Add test for new dummy base_estimator

Check that the dummy estimator does not fit models and always
reutrns a random number.

* Add comment in changelog