Set log-scale in plots if `prior=log-uniform`.
`plot_objective` and `plot_evaluations` change axis scales to log-scale
depending on the priors of dimensions.

For `plot_evaluations`, this means that binning has to be
log-transformed as well to create meaningful plots.

Instances of `class Integer(Dimension)` didn't have a `prior` attribute.
Now, there's a fallback, which is `"uniform"`, as a class attribute of
the super-class `Dimension`.