Fix compatibility with sklearn
Doing *anything* in the `__init__` method of a sklearn estimator will most of the time break his compatibility with sklearn's clone, get_params and set_params, or the ability to fit it several times. It is in fact the case here, it would break if used in a Pipeline or other meta-objects in sklearn.

This PR is a quick proposal to fix the compatibility issues.